+++ benihora-vst/src/benihora_managed/pid_controller.rs
@@ -1,26 +1,26 @@
 use serde::{Deserialize, Serialize};
 
 pub struct PIDController {
-    pub(super) dtime: f64,
-    integral: f64,
-    last: f64,
+    pub(super) dtime: f32,
+    integral: f32,
+    last: f32,
 }
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize)]
 pub struct PIDParam {
-    pub kp: f64,
-    pub ki: f64,
-    pub kd: f64,
+    pub kp: f32,
+    pub ki: f32,
+    pub kd: f32,
 }
 
 impl PIDParam {
-    pub fn new(kp: f64, ki: f64, kd: f64) -> Self {
+    pub fn new(kp: f32, ki: f32, kd: f32) -> Self {
         Self { kp, ki, kd }
     }
 }
 
 impl PIDController {
-    pub fn new(sample_rate: f64) -> Self {
+    pub fn new(sample_rate: f32) -> Self {
         Self {
             dtime: 1.0 / sample_rate,
             integral: 0.0,
@@ -28,7 +28,7 @@ impl PIDController {
         }
     }
 
-    pub fn process(&mut self, pid: &PIDParam, x: f64) -> f64 {
+    pub fn process(&mut self, pid: &PIDParam, x: f32) -> f32 {
         let d = x - self.last;
         self.integral = self.integral + x * self.dtime;
         let y = (x * pid.kp + self.integral * pid.ki) * self.dtime + d * pid.kd;
